% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaup.R
\name{beaup}
\alias{beaup}
\title{Petit p bien presenté}
\usage{
beaup(varp, affp = FALSE)
}
\arguments{
\item{varp}{p.value}

\item{affp}{if TRUE the output will be like "p = 0.005" else simply "0,005" (default : FALSE)}
}
\value{
beaup , chaine de caractère
}
\description{
Petit p bien presenté
}
\examples{
pp <- cor.test(iris$Sepal.Length,iris$Sepal.Width)
beaup(pp$p.value)


}
