% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnettgent.R
\name{dunnettgent}
\alias{dunnettgent}
\title{dunnettgent}
\usage{
dunnettgent(dfx, vin, vout, success, delta, corcont = FALSE)
}
\arguments{
\item{dfx}{a data.frame or a tibble}

\item{vin}{binary variable that may affect the result}

\item{vout}{binary variable success/fail}

\item{success}{leel of vout for success}

\item{delta}{non-inferiority limit}

\item{corcont}{corrected for continuity ? (default : FALSE)}
}
\value{
khi2 and p-value
}
\description{
Khi2 modified by Dunnett & Gent for non-inferiority tests, very often used in clinical studies.
}
\examples{
df <- dplyr::tibble(trait = c("a","b","a"), alive = c("yes", "yes", "no"))
dunnettgent(dfx = df, vin = trait, vout = alive, success = "yes", delta = 0.1, corcont = TRUE)

}
